# docs/swagger.yaml
definitions:
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /container/fs/{id}:
    post:
      description: |-
        Forwards MCP JSON-RPC requests to the MCP server inside the container.
        Required:
        - container task is running
        - container has data mount (default /data) bound to <data_root>/users/<user_id>
        - container image contains the "mcp" binary
        Auth: Bearer JWT is used to determine user_id (sub or user_id).
        Paths must be relative (no leading slash) and must not contain "..".

        Example: tools/list
        {"jsonrpc":"2.0","id":1,"method":"tools/list"}

        Example: tools/call (fs.read)
        {"jsonrpc":"2.0","id":2,"method":"tools/call","params":{"name":"fs.read","arguments":{"path":"notes.txt"}}}
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Container ID
        in: path
        name: id
        required: true
        type: string
      - description: JSON-RPC request
        in: body
        name: payload
        required: true
        schema:
          type: object
      responses:
        "200":
          description: 'JSON-RPC response: {jsonrpc,id,result|error}'
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: MCP filesystem tools (JSON-RPC)
      tags:
      - containerd
...
swagger: "2.0"